Updating to version 2_0_0-RC1
This is a large update with many breaking changes. It includes a complete re-architecture
and change to the public API of the methods. Hence the major version change. This is in
response to more experience using the types in multiple situations. The previous version
was more verbose and required more knowledge on the part of the end user than was desirable
to understand the ins and outs of how waiting for conditions was done. The change in this
update vastly simplifies the library and makes it more customizable, easily, for end users.